fix: cable energy network distribution, block logic audit, PMD 423->0
- Energy: distribute over the whole cable network once per tick
  (pull from pure sources, redistribute buffers, push to sinks) instead
  of pushing into every neighbor incl. adjacent cables (ping-pong)
- Speaker: replace texture/model with OpenComputers Charger look
- Fix 5 bugs found in block/monitor audit: framebuffer dirtyLines never
  cleared (perpetual re-encode), network switch load crash on empty adj
  list, empty-frame packet spam on null network interfaces, PCI card
  cage mount/energy state, bundled redstone output/input mirroring
- Clean up PMD: 423 violations to 0 (hoisted allocations, extracted
  helpers for high-complexity methods, NOPMD markers with justification,
  renamed ToInt/Copy/ToClient/ToServer, getter renames left as public
  API churn with inline NOPMD)
- Monitor: replace OBJ model with programmatic model, multiblock
  merge/break/repartition logic

+++ src/main/java/li/cil/oc2/common/vm/terminal/fonts/FontAtlas.java
@@ -82,10 +82,10 @@ public class FontAtlas {
         }
 
         // Calculate the UV coordinates for this glyph
-        float uStart = (float) currentX * (1f / atlasWidth);
-        float vStart = (float) currentY * (1f / atlasHeight);
-        float uEnd = (float) (currentX + glyph.image.getWidth() + 1) * (1f / atlasWidth);
-        float vEnd = (float) (currentY + glyph.image.getHeight() + 1) * (1f / atlasHeight);
+        float uStart = currentX * (1f / atlasWidth);
+        float vStart = currentY * (1f / atlasHeight);
+        float uEnd = (currentX + glyph.image.getWidth() + 1) * (1f / atlasWidth);
+        float vEnd = (currentY + glyph.image.getHeight() + 1) * (1f / atlasHeight);
 
         glyph.setUV(uStart, vStart, uEnd, vEnd);
